KE-9472 change query history storage from influxDB to RDBMS

+++ src/server-base/src/main/java/io/kyligence/kap/rest/response/QueryStatisticsResponse.java
@@ -32,10 +32,10 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class QueryStatisticsResponse {
-    private int count;
+    private long count;
     private double mean;
 
-    public QueryStatisticsResponse(int count, double mean) {
+    public QueryStatisticsResponse(long count, double mean) {
         this.count = count;
         this.mean = mean;
     }
